Fixes captive.rs (#5)
* Fixed xSemaphoreTake and StackOverflow errors. Added a captive portal example.
It isn't possible to move an UDP socket accross thread bounds on esp32.
So it has to be created in the same thread as it's polled.
Additionally more stack memory is needed to run the server.
Changed info! logs to debug! logs.

* Minor Changes
Removed domain from the defaults.
Removed unnecessary domain feature.